+++ dataservices/admin.py
@@ -25,16 +25,30 @@ class EaseOfDoingBusinessAdmin(admin.ModelAdmin):
     formfield_overrides = {TextField: {'widget': forms.TextInput}}
 
     search_fields = (
-        'country',
+        'country__name',
         'year',
         'value',
     )
 
     list_display = (
         'country',
+        'get_iso2',
+        'get_iso3',
         'year',
         'value',
     )
+
+    def get_iso2(self, obj):
+        return obj.country.iso2
+
+    get_iso2.admin_order_field = 'country'
+    get_iso2.short_description = 'iso2'
+
+    def get_iso3(self, obj):
+        return obj.country.iso3
+
+    get_iso3.admin_order_field = 'country'
+    get_iso3.short_description = 'iso3'
 
 
 @admin.register(models.CorruptionPerceptionsIndex)
@@ -96,16 +110,30 @@ class InternetUsageAdmin(admin.ModelAdmin):
     formfield_overrides = {TextField: {'widget': forms.TextInput}}
 
     search_fields = (
-        'country',
+        'country__name',
         'year',
         'value',
     )
 
     list_display = (
         'country',
+        'get_iso2',
+        'get_iso3',
         'year',
         'value',
     )
+
+    def get_iso2(self, obj):
+        return obj.country.iso2
+
+    get_iso2.admin_order_field = 'country'
+    get_iso2.short_description = 'iso2'
+
+    def get_iso3(self, obj):
+        return obj.country.iso3
+
+    get_iso3.admin_order_field = 'country'
+    get_iso3.short_description = 'iso3'
 
 
 @admin.register(models.ConsumerPriceIndex)
@@ -113,16 +141,30 @@ class ConsumerPriceIndexAdmin(admin.ModelAdmin):
     formfield_overrides = {TextField: {'widget': forms.TextInput}}
 
     search_fields = (
-        'country',
+        'country__name',
         'year',
         'value',
     )
 
     list_display = (
         'country',
+        'get_iso2',
+        'get_iso3',
         'year',
         'value',
     )
+
+    def get_iso2(self, obj):
+        return obj.country.iso2
+
+    get_iso2.admin_order_field = 'country'
+    get_iso2.short_description = 'iso2'
+
+    def get_iso3(self, obj):
+        return obj.country.iso3
+
+    get_iso3.admin_order_field = 'country'
+    get_iso3.short_description = 'iso3'
 
 
 @admin.register(models.Country)
@@ -153,9 +195,27 @@ class CountryAdmin(admin.ModelAdmin):
 class GDPPerCapitaAdmin(admin.ModelAdmin):
     list_display = (
         'country',
+        'get_iso2',
+        'get_iso3',
         'year',
         'value',
     )
+    search_fields = (
+        'country__name',
+        'year',
+    )
+
+    def get_iso2(self, obj):
+        return obj.country.iso2
+
+    get_iso2.admin_order_field = 'country'
+    get_iso2.short_description = 'iso2'
+
+    def get_iso3(self, obj):
+        return obj.country.iso3
+
+    get_iso3.admin_order_field = 'country'
+    get_iso3.short_description = 'iso3'
 
     class Meta:
         model = models.GDPPerCapita
@@ -171,9 +231,27 @@ class SuggestedCountryAdmin(admin.ModelAdmin):
 class IncomeAdmin(admin.ModelAdmin):
     list_display = (
         'country',
+        'get_iso2',
+        'get_iso3',
         'year',
         'value',
     )
+    search_fields = (
+        'country__name',
+        'year',
+    )
+
+    def get_iso2(self, obj):
+        return obj.country.iso2
+
+    get_iso2.admin_order_field = 'country'
+    get_iso2.short_description = 'iso2'
+
+    def get_iso3(self, obj):
+        return obj.country.iso3
+
+    get_iso3.admin_order_field = 'country'
+    get_iso3.short_description = 'iso3'
 
 
 @admin.register(models.RuleOfLaw)